+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/orbax_layout.py
@@ -20,14 +20,16 @@ from typing import Any, Awaitable
 
 from absl import logging
 from orbax.checkpoint._src import asyncio_utils
+from orbax.checkpoint._src.metadata import step_metadata_serialization
 from orbax.checkpoint._src.multihost import multihost
 from orbax.checkpoint._src.path import async_path
 from orbax.checkpoint._src.path import temporary_paths
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 from orbax.checkpoint.experimental.v1._src.handlers import registration
 from orbax.checkpoint.experimental.v1._src.handlers import resolution as handler_resolution
+from orbax.checkpoint.experimental.v1._src.handlers import types as handler_types
 from orbax.checkpoint.experimental.v1._src.layout import checkpoint_layout
-from orbax.checkpoint.experimental.v1._src.loading import v0_compatibility
+from orbax.checkpoint.experimental.v1._src.metadata import serialization as metadata_serialization
 from orbax.checkpoint.experimental.v1._src.metadata import types as metadata_types
 from orbax.checkpoint.experimental.v1._src.path import types as path_types
 from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
@@ -39,8 +41,12 @@ class CheckpointVersion(enum.Enum):
 
 
 InvalidLayoutError = checkpoint_layout.InvalidLayoutError
+NoEntryError = registration.NoEntryError
 Path = path_types.Path
 CheckpointLayout = checkpoint_layout.CheckpointLayout
+InternalCheckpointMetadata = (
+    step_metadata_serialization.InternalCheckpointMetadata
+)
 
 PYTREE_METADATA_FILE = "_METADATA"
 ORBAX_CHECKPOINT_INDICATOR_FILE = "orbax.checkpoint"
@@ -149,6 +155,19 @@ async def _create_orbax_identifier_file(
     )
 
 
+async def read_checkpoint_metadata(
+    directory: path_types.Path,
+) -> InternalCheckpointMetadata:
+  """Returns the step metadata for a given path."""
+  serialized_metadata = (
+      await metadata_serialization.read(
+          metadata_serialization.checkpoint_metadata_file_path(directory)
+      )
+      or {}
+  )
+  return InternalCheckpointMetadata.deserialize(serialized_metadata)
+
+
 class OrbaxLayout(CheckpointLayout):
   """OrbaxLayout.
 
@@ -172,13 +191,21 @@ class OrbaxLayout(CheckpointLayout):
       self, path: Path
   ) -> metadata_types.CheckpointMetadata[dict[str, Any]]:
     """Returns the metadata describing the Orbax checkpoint."""
-    # Uses the v0 checkpointer to get v0 StepMetadata
-    checkpointer, _ = v0_compatibility.get_v0_checkpointer_and_args(
-        path, None, context=context_lib.get_context()
+    checkpoint_metadata = await read_checkpoint_metadata(path)
+    available_checkpointable_names = await _existing_checkpointable_names(path)
+    abstract_checkpointables = {
+        name: None for name in available_checkpointable_names
+    }
+    handlers_for_load = await handler_resolution.get_handler_for_load_metadata(
+        self._handler_registry,
+        abstract_checkpointables,
+        checkpoint_metadata,
     )
-    step_metadata = checkpointer.metadata(path)
 
-    item_metadata = {k: v for k, v in step_metadata.item_metadata.items()}
+    item_metadata = await _get_checkpointables_metadata(
+        path, abstract_checkpointables, handlers_for_load
+    )
+
     # Exclude `metrics` if present. This is relevant only for
     # `training.Checkpointer`, and is separately added to the
     # `training.CheckpointMetadata` object.
@@ -186,9 +213,9 @@ class OrbaxLayout(CheckpointLayout):
 
     return metadata_types.CheckpointMetadata[dict[str, Any]](
         metadata=item_metadata,
-        init_timestamp_nsecs=step_metadata.init_timestamp_nsecs,
-        commit_timestamp_nsecs=step_metadata.commit_timestamp_nsecs,
-        custom_metadata=step_metadata.custom_metadata,
+        init_timestamp_nsecs=checkpoint_metadata.init_timestamp_nsecs,
+        commit_timestamp_nsecs=checkpoint_metadata.commit_timestamp_nsecs,
+        custom_metadata=checkpoint_metadata.custom_metadata,
     )
 
   async def _validate_pytree(self, path: Path, checkpointable_name: str | None):
@@ -332,10 +359,19 @@ class OrbaxLayout(CheckpointLayout):
     Returns:
       An awaitable containing the loaded pytree.
     """
-    load_awaitable = await self.load_checkpointables(
-        path, {checkpointable_name: abstract_pytree}
+    checkpoint_metadata = await read_checkpoint_metadata(path)
+    handlers_for_load = await handler_resolution.get_handlers_for_load(
+        self._handler_registry,
+        {checkpointable_name: abstract_pytree},
+        checkpoint_metadata,
     )
-    return load_awaitable
+
+    handler_for_load = handlers_for_load[checkpointable_name]
+    result = await handler_for_load.load(
+        path / checkpointable_name,
+        abstract_pytree,
+    )
+    return result
 
   async def load_checkpointables(
       self,
@@ -352,30 +388,52 @@ class OrbaxLayout(CheckpointLayout):
       An awaitable containing the loaded checkpointables.
 
     Raises:
-      KeyError: If any of the specified checkpointable names are not found in
-      the checkpoint.
+      KeyError: If an empty checkpoint is attempted to be loaded or there are
+        no valid checkpointables to load from the user passed
+        `abstract_checkpointables`.
     """
-    abstract_checkpointables = abstract_checkpointables or {}
-    handlers_for_load = await handler_resolution.get_handlers_for_load(
-        path, self._handler_registry, abstract_checkpointables
-    )
-    existing_checkpointable_names = await _existing_checkpointable_names(path)
+    available_checkpointables = await _existing_checkpointable_names(path)
     if not abstract_checkpointables:
+      # Default to loading all checkpointables (subdirectories within current
+      # checkpoint directory)
       abstract_checkpointables = {
           name: None
-          for name in handlers_for_load.keys()
+          for name in available_checkpointables
           if name not in checkpoint_layout.RESERVED_CHECKPOINTABLE_KEYS
-          and name in existing_checkpointable_names
       }
+
     if any(
-        name not in existing_checkpointable_names
+        name not in available_checkpointables
         for name in abstract_checkpointables.keys()
     ):
       raise KeyError(
-          f"Requested checkpointables: {abstract_checkpointables.keys()} for"
+          f" Requested checkpointables: {abstract_checkpointables.keys()} for"
           " loading were not found in the checkpoint. Available"
-          f" checkpointables: {existing_checkpointable_names}"
+          f" checkpointables: {available_checkpointables}"
       )
+
+    # Read checkpoint metadata and resolve handlers for loading.
+    checkpoint_metadata = await read_checkpoint_metadata(path)
+    # TODO(b/484400394): Find a better way to inform the user that they need
+    # to use load_pytree(..., checkpointable_name=None) when item_handlers is
+    # a str. An idea is to create a seperate validate_checkpointables method
+    # and we can read in checkpoint metadata at validation time for both
+    # validate_pytree and validate_checkpointables operations and warn the user
+    # know if they are trying to load a composite checkpoint by calling
+    # load_pytree(checkpointable_name=None) or trying to load a composite
+    # checkpoint as a pytree checkpoint respectively.
+    if isinstance(checkpoint_metadata.item_handlers, str):
+      logging.warning(
+          "Checkpoint looks like a legacy V0 checkpoint. This is only"
+          " supported for legacy V0 checkpoints. If you intended to load a"
+          " pytree checkpoint from the given path, then please consider using"
+          " `loading.load_pytree(..., checkpointable_name=None)` instead."
+      )
+    handlers_for_load = await handler_resolution.get_handlers_for_load(
+        self._handler_registry,
+        abstract_checkpointables,
+        checkpoint_metadata,
+    )
 
     load_ops = []
     for (
@@ -392,19 +450,9 @@ class OrbaxLayout(CheckpointLayout):
     load_awaitables = await asyncio.gather(*load_ops)
 
     async def _run_background() -> dict[str, Any]:
-      loaded_checkpointables = []
-      # TODO(b/398249409) Cannot use asyncio.gather because asyncio.run
-      # is used in underlying implementation.
-      for a in load_awaitables:
-        loaded = await a
-        loaded_checkpointables.append(loaded)
-      return {
-          checkpointable_name: loaded
-          for checkpointable_name, loaded in zip(
-              abstract_checkpointables.keys(),
-              loaded_checkpointables,
-          )
-      }
+      return await _process_load_awaitables(
+          abstract_checkpointables, load_awaitables, handlers_for_load
+      )
 
     return _run_background()
 
@@ -448,3 +496,83 @@ class OrbaxLayout(CheckpointLayout):
       )
 
     return _run_background()
+
+
+async def _get_checkpointables_metadata(
+    path: Path,
+    abstract_checkpointables: dict[str, Any],
+    handlers: dict[str, handler_types.CheckpointableHandler | None],
+) -> dict[str, Any | None]:
+  """Retrieves metadata for each checkpointable item.
+
+  For each checkpointable, the metadata is retrieved by calling the
+  `metadata` method on the handler for the checkpointable. If the handler is
+  None or if we encounter an error while loading the metadata, then the metadata
+  for the checkpointable is set to None.
+
+  Args:
+    path: The path to the checkpoint directory.
+    abstract_checkpointables: The abstract checkpointables to load.
+    handlers: The handlers for the checkpointables.
+
+  Returns:
+    A mapping from checkpointable name to metadata.
+  """
+  item_metadata = {}
+  for checkpointable_name in abstract_checkpointables.keys():
+    handler = handlers[checkpointable_name]
+    if handler is None:
+      item_metadata[checkpointable_name] = None
+      continue
+    try:
+      checkpointable_metadata = await handler.metadata(
+          path / checkpointable_name
+      )
+    except BaseException as e:  # pylint:disable=broad-exception-caught
+      logging.warning(
+          "Failed to get metadata for checkpointable: %s due to incompatible"
+          " handler: %s for abstract checkpointable: %s. Defaulting return"
+          " value to None. \nError details: %s",
+          checkpointable_name,
+          handler,
+          abstract_checkpointables[checkpointable_name],
+          e,
+      )
+      checkpointable_metadata = None
+    item_metadata[checkpointable_name] = checkpointable_metadata
+  return item_metadata
+
+
+async def _process_load_awaitables(
+    checkpointables_to_load: dict[str, Any],
+    load_awaitables: list[Any],
+    handlers: dict[str, handler_types.CheckpointableHandler],
+) -> dict[str, Any]:
+  """Processes the awaitables after we initiate all load operations.
+
+  For each checkpointable, the load operation is processed by calling the
+  `load` method on the handler for the checkpointable.
+
+  Args:
+    checkpointables_to_load: The checkpointables to load.
+    load_awaitables: The awaitables for the load operations.
+    handlers: The handlers used to load the checkpointables.
+
+  Returns:
+    A mapping from checkpointable name to loaded checkpointable.
+  """
+  loaded_checkpointables = {}
+  # TODO(b/398249409) Cannot use asyncio.gather because asyncio.run
+  # is used in underlying implementation.
+  for name, a in zip(checkpointables_to_load.keys(), load_awaitables):
+    try:
+      loaded = await a
+      loaded_checkpointables[name] = loaded
+    except BaseException as e:  # pylint:disable=broad-exception-caught
+      raise registration.NoEntryError(
+          f"Failed to load checkpointable: {name} due to incompatible handler:"
+          f" {handlers[name]} for abstract checkpointable:"
+          f" {checkpointables_to_load[name]}. \nError details: {e}"
+      ) from e
+
+  return loaded_checkpointables